Update tests for extracted <PopupButton>
